Configure lib. Added first models

+++ lib/src/main/java/com/melgarejojunior/calculator/lib/Polygon.kt
@@ -1,6 +1,6 @@
 package com.melgarejojunior.calculator.lib
 
 interface Polygon {
-    fun area(): Float
-    fun perimeter(): Float
+    fun area(): Double
+    fun perimeter(): Double
 }
